breaking out tab panels
