+++ src/components/Tabs/Tabs.Overview.stories.mdx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 import { Tabs } from './Tabs';
-import { TabPanels } from './TabPanels';
+import { TabPanels } from '../TabPanels/TabPanels';
 import { Box } from '../Box/Box';
 import { Button } from '../Button/Button';
