+++ src/components/Tabs/Tabs.test.tsx
@@ -128,25 +128,4 @@ describe('Tabs', () => {
       expect(mockedOnClick).toHaveBeenCalledTimes(1);
     });
   });
-
-  describe('Tab Panels', () => {
-    test('it shows the correct panel based on the selected tab', () => {
-      render(
-        <TabPanels value={1}>
-          <div>
-            tab 1
-          </div>
-          <div>
-            tab 2
-          </div>
-        </TabPanels>,
-      );
-
-      const tabOne = screen.queryByText('tab 1');
-      const tabTwo = screen.queryByText('tab 2');
-
-      expect(tabOne).not.toBeInTheDocument();
-      expect(tabTwo).toBeInTheDocument();
-    });
-  });
 });
