+++ src/components/TabsSlider/TabsSlider.Overview.stories.mdx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 import { TabsSlider } from './TabsSlider';
-import { TabPanels } from './TabPanels';
+import { TabPanels } from '../TabPanels/TabPanels';
 import { Box } from '../Box/Box';
 import { Button } from '../Button/Button';
