+++ src/components/index.ts
@@ -34,9 +34,9 @@ export * from './ResponsiveProvider/ResponsiveProvider';
 export * from './SelectInput/SelectInput';
 export * from './SelectInputNative/SelectInputNative';
 export * from './Spinner/Spinner';
+export * from './TabPanels/TabPanels';
 export * from './Tabs/Tabs';
-export * from './Tabs/TabItem';
-export * from './Tabs/TabPanels';
+export * from './TabsSlider/TabsSlider';
 export * from './Table/Table';
 export * from './TextInput/TextInput';
 export * from './TextareaInput/TextareaInput';
